+++ federated_aws_cli/.editorconfig
