+++ federated_aws_cli/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Top-level package for federated-boto."""
+"""Top-level package for federated-aws-cli."""
 
 __author__ = """Mozilla Infosec"""
 __email__ = 'infosec@mozilla.com'
