+++ setup.py
@@ -38,19 +38,19 @@ setup(
     description="CLI application that handled federated authentication for AWS users",
     entry_points={
         'console_scripts': [
-            'federated_boto=federated_boto.cli:main',
+            'federated_aws_cli=federated_aws_cli.cli:main',
         ],
     },
     install_requires=requirements,
     long_description=readme + '\n\n' + history,
     include_package_data=True,
-    keywords='federated_boto',
-    name='federated_boto',
-    packages=find_packages(include=['federated_boto']),
+    keywords='federated_aws_cli',
+    name='federated_aws_cli',
+    packages=find_packages(include=['federated_aws_cli']),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
-    url='https://github.com/mozilla-iam/federated_boto',
+    url='https://github.com/mozilla-iam/federated_aws_cli',
     version='0.0.1',
     zip_safe=False,
 )
