+++ tox.ini
@@ -11,7 +11,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 federated_boto
+commands = flake8 federated_aws_cli
 
 [testenv]
 setenv =
